EditorTools: Object Name, Texture & toggle changes now saving.
When changing an objects name, texture or toggle state in the debug
menu, the changes can now be saved and loaded.
Added filename string to model component. When changing the name of the
object, we are changing the id of the object. The file name represents
the obj name.
Updated Level1.xml to represent any editor tools requirements.
Removed redundant object name check when saving a file, as objects can
be renamed.