EditorTools: DebugUtils class
A class to be used from any file that provides usefull debugging
functions to the screen.
First example use:
A popup can be used when a error occurred (instead of console) or a
success.
Or perhaps for any bool case where on screen information may be useful.
DebugUtils::getInstance()->popup("FAIL","Failed to load file, filename
!= to levelname");