UI: Fixed gamescreen no render issue, Clean + debugmenu integration
UI now rendering on current thread as suggested by DOM. VAOs are not
shared so one is created on first render.
Added UIManager to handle the debug menu for UI across screens. ( the
format is the same, the UIelements in the vector are different.)
Debug menu integration for UI, allows for changing the transform of the
UI.
UI read in from file and then into a Transform.
TODO: Save to file, change texture and name for UI in debugmenu