UI: Create TextBox, file save optimisations
You can now create a text box via the debug menu
You can now change font and colour of text via debug menu.
WARNING: Changing of font is untested.
Optimised FileSaver.cpp to with add and update vec3 function. Every time
we want to update or add a vec3 to a file, we can reuse the functions.
Lots of code deleted as a result.
TODO:
-Make button a UIElement.
-Optimise file saver for add / update vec4. This is non priority.