UI: Can now add a button to an existing UIElement
You can now add a button to a UIElement via the debug menu.
Remember to add .lua to the script name. Will probably change this to a
listbox for error checking.
Extra check added if a uiElement has no button and when the loaded
button has no string content in any field.